Split template rendering and delegating node rendering
This step makes it easier to add your own template engine, all
node renderering is now done by the node renderers.
